Make default formatter internal and update formatter interface
