+++ detect.go
@@ -26,6 +26,7 @@ import (
 	"github.com/Masterminds/semver"
 
 	"github.com/buildpacks/libcnb/internal"
+	"github.com/buildpacks/libcnb/log"
 )
 
 // DetectContext contains the inputs to detection.
@@ -39,7 +40,7 @@ type DetectContext struct {
 	Buildpack Buildpack
 
 	// Logger is the way to write messages to the end user
-	Logger Logger
+	Logger log.Logger
 
 	// Platform is the contents of the platform.
 	Platform Platform
@@ -77,10 +78,7 @@ func Detect(detect DetectFunc, config Config) {
 	}
 
 	if config.logger.IsDebugEnabled() {
-		if err := internal.NewDirectoryContentsWriter(
-			ctx.ApplicationPath,
-			config.dirContentFormatter,
-			config.logger.DebugWriter()).Write("Application contents"); err != nil {
+		if err := config.contentWriter.Write("Application contents", ctx.ApplicationPath); err != nil {
 			config.logger.Debugf("unable to write application contents\n%w", err)
 		}
 	}
@@ -93,10 +91,7 @@ func Detect(detect DetectFunc, config Config) {
 	}
 
 	if config.logger.IsDebugEnabled() {
-		if err := internal.NewDirectoryContentsWriter(
-			ctx.Buildpack.Path,
-			config.dirContentFormatter,
-			config.logger.DebugWriter()).Write("Buildpack contents"); err != nil {
+		if err := config.contentWriter.Write("Buildpack contents", ctx.Buildpack.Path); err != nil {
 			config.logger.Debugf("unable to write buildpack contents\n%w", err)
 		}
 	}
@@ -140,10 +135,7 @@ func Detect(detect DetectFunc, config Config) {
 	}
 
 	if config.logger.IsDebugEnabled() {
-		if err := internal.NewDirectoryContentsWriter(
-			ctx.Platform.Path,
-			config.dirContentFormatter,
-			config.logger.DebugWriter()).Write("Platform contents"); err != nil {
+		if err := config.contentWriter.Write("Platform contents", ctx.Platform.Path); err != nil {
 			config.logger.Debugf("unable to write platform contents\n%w", err)
 		}
 	}
