+++ log/init_test.go
@@ -26,6 +26,5 @@ import (
 func TestUnit(t *testing.T) {
 	suite := spec.New("libcnb/log", spec.Report(report.Terminal{}))
 	suite("PlainLogger", testLogger)
-	suite("PlainFormatter", testFormatters)
 	suite.Run(t)
 }
